requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.1.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
frozenlist==1.4.1
google-api-core==2.14.0
google-api-python-client==2.109.0
google-auth==2.24.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
magic-filter==1.0.12
multidict==6.0.5
numpy==1.26.2
oauth2client==4.1.3
oauthlib==3.2.2
outcome==1.3.0.post0
pandas==2.1.3
protobuf==4.25.1
pyaes==1.6.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
pyparsing==3.1.1
Pyrogram==2.0.106
PySocks==1.7.1
pyTelegramBotAPI==4.14.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
TgCrypto==1.2.5
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.18
wsproto==1.2.0
yarl==1.9.4
